test: Negative retries throws error
