+++ edgar_tool/__init__.py
@@ -1,4 +1,9 @@
+from edgar_tool.cli import app
 from edgar_tool.rss import fetch_rss_feed
 from edgar_tool.text_search import EdgarTextSearcher
 
-__all__ = ["EdgarTextSearcher", "fetch_rss_feed"]
+__all__ = [
+    "app",
+    "EdgarTextSearcher",
+    "fetch_rss_feed",
+]
