+++ tests/test_cli.py
@@ -1,5 +1,11 @@
 import subprocess
 
+from typer.testing import CliRunner
+
+import edgar_tool
+
+runner = CliRunner()
+
 
 def test_cli_should_return_help_string_when_passed_no_args():
     """Tests that running edgar-tool without any arguments returns the CLI's help string and 0 exit code."""
@@ -21,6 +27,8 @@ Usage: edgar-tool [OPTIONS] COMMAND [ARGS]...
 ╰──────────────────────────────────────────────────────────────────────────────╯
 
 """
+    # Split the lines so we can strip off any trailing whitespace characters.
+    # The subprocess output puts spaces after [ARGS]...
     expected_lines = [line.strip() for line in expected_output.splitlines()]
 
     # WHEN
@@ -30,3 +38,25 @@ Usage: edgar-tool [OPTIONS] COMMAND [ARGS]...
     # THEN
     assert result.returncode == 0
     assert actual_lines == expected_lines
+
+
+def test_text_search_negative_retries():
+    """
+    Tests that passing a negative value for --retries throws an error.
+    """
+    # GIVEN/WHEN
+    result = runner.invoke(
+        edgar_tool.app,
+        [
+            "text-search",
+            "example",  # At least one search term is required
+            "--retries",
+            "-1",
+        ],
+    )
+    # THEN
+    assert result.exit_code != 0
+    assert (
+        "Invalid value for '--retries' / '-r': -1 is not in the range x>=0."
+        in result.output
+    )
